Carousel is now fully controlled by the straipi backend, all content changes and the carousel appearing on the chosen pages is administered there
