+++ components/sections/carousel/index.js
@@ -1,0 +1,3 @@
+import EmblaCarousel from "./carousel"
+
+export default EmblaCarousel
